refactor: split config classes
