+++ packages/portal/tests/unit/plugins/europeana/annotation.spec.js
@@ -1,10 +1,8 @@
 import nock from 'nock';
 
-import annotation, { BASE_URL } from '@/plugins/europeana/annotation';
+import EuropeanaAnnotationApi from '@/plugins/europeana/annotation';
 
-const store = { state: { apis: { urls: { annotation: BASE_URL } } } };
-
-describe('plugins/europeana/entity', () => {
+describe('plugins/europeana/annotation', () => {
   afterEach(() => {
     nock.cleanAll();
   });
@@ -12,14 +10,15 @@ describe('plugins/europeana/entity', () => {
   describe('search', () => {
     it('searches Annotation API', async() => {
       const apiQuery = '*:*';
-      nock(BASE_URL)
+      nock(EuropeanaAnnotationApi.BASE_URL)
         .get('/search')
         .query(query => {
           return query.query === apiQuery;
         })
         .reply(200, {});
+      const api = new EuropeanaAnnotationApi;
 
-      await annotation({ store }).search({ query: apiQuery });
+      await api.search({ query: apiQuery });
 
       expect(nock.isDone()).toBe(true);
     });
